Updates for docker builds (#1)
* Update Makefiles in PBC and src for gcc 6XX name mangling with older libraries

* Minor correction to Dockerfile

* Delete dlg-space.df

* Update Makefiles to account for gcc 6 name mangling

* Correct working directory

* Test patch dockerfile

* Corrected annoying warnings

* Tested under Ubuntu 14.04 / 16.04

* Returned Dockerfile to pull from master

+++ src/buddy.cc
@@ -217,7 +217,7 @@ int Buddy::read_messagestr(string &msgstr) const
     unsigned char header[headerLength];
 
     size_t res = read_record(header, headerLength);
-    if (res < headerLength) return -1;
+    if (res < (unsigned int)headerLength) return -1;
 
     int lengthFieldStart = IDLength + typeLength;
     unsigned int len = (header[lengthFieldStart] << 24) +
